feat: layout, login, home pages lazy loading
